Add files via upload
Developed a number of Use-case diagrams to explain the processes done by the mail room Actors (operators). Use case diagrams model the functionality of a system using actors and use cases. Use cases are a set of actions, services, and functions that the system needs to perform. The actor is an entity that interacts with the system. A user is the best example of an actor. An actor is an entity that initiates the use case from outside the scope of a use case. it describes the user's interaction with the system that shows the relationship between the user and the different use cases in which the user is involved.